// src/store/types.ts
export type NodeType = 'corner' | 'smooth' | 'offcurve' | 'qcurve'

export interface PathNode {
  id: string
  x: number
  y: number
  type: NodeType
}

export interface PathData {
  id: string
  nodes: PathNode[]
  closed: boolean
}